split logic of validation/inputs.py and validation/errors_tracker.py
