+++ src/python/impactx/dashboard/Input/validation/__init__.py
@@ -1,0 +1,7 @@
+from .inputs import DashboardValidation
+from .errors_tracker import ErrorsTracker
+
+__all__ = [
+    "DashboardValidation",
+    "ErrorsTracker",
+]
